Split path and surface stats code into separate files
